# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MDDB API
  description: |
    MDDB (Markdown Database) is a high-performance, version-controlled markdown database with triple protocol support.

    ## Features
    - Store and manage markdown documents with metadata
    - Full revision history for all documents
    - Multi-language support
    - Fast metadata-based search
    - **Vector search** with embedding providers (OpenAI, Cohere, Voyage AI, Ollama)
    - **Full-text search** with TF-IDF, BM25, BM25F, PMISparse scoring, 7 search modes (simple, boolean, phrase, wildcard, proximity, range, fuzzy), and multi-language stemming (18 languages)
    - **Hybrid search** combining FTS + vector search with alpha blending or RRF fusion
    - **Webhooks** for real-time event notifications
    - **TTL** (time-to-live) support for automatic document expiration
    - **JSON Schema validation** for document metadata
    - **URL import** with YAML frontmatter parsing
    - **Authentication** - JWT tokens and API keys
    - **Authorization** - Collection-level RBAC (Read/Write/Admin)
    - **Memory RAG** - Conversational memory system with session management, message storage, semantic/hybrid recall, and summarization
    - **User Management** - Multi-user with admin roles
    - **Group Permissions** - Organize users into groups
    - Template variable substitution
    - Export capabilities (NDJSON, ZIP)
    - Backup and restore operations
...

    ## Protocols
    - **HTTP/JSON** - Port 11023 (documented here)
...
servers:
...
paths:
...
  /v1/auth/api-keys:
    get:
      tags:
        - Authentication
      summary: List API keys
      description: List all API keys for the authenticated user (key values are not
        included, only hashes).
      operationId: authListAPIKeys
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAPIKeysResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/auth/users:
    get:
      tags:
        - Users
      summary: List all users
      description: List all user accounts with their details. Requires admin authentication.
      operationId: authListUsers
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UsersListResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /v1/auth/groups:
    get:
      tags:
        - Groups
      summary: List all groups
      description: List all user groups. Requires admin authentication.
      operationId: authListGroups
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of groups
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupsListResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags:
        - Groups
      summary: Create group
      description: Create a new user group. Requires admin authentication.
      operationId: authCreateGroup
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
      responses:
        "201":
          description: Group created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
        "400":
          description: Invalid request (missing name or group already exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: document not found
...
    ListAPIKeysResponse:
      type: object
      properties:
        keys:
          type: array
          items:
            type: object
            properties:
              keyHash:
                type: string
                description: SHA256 hash of the API key (for deletion)
                example: a1b2c3d4e5f6...
              description:
                type: string
                description: API key description
                example: CI/CD pipeline key
              createdAt:
                type: integer
                format: int64
                description: Creation timestamp (Unix seconds)
                example: 1704844800
              expiresAt:
                type: integer
                format: int64
                description: Expiration timestamp (Unix seconds, 0 = never)
                example: 0
...
    UsersListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            type: object
            properties:
              username:
                type: string
                description: Username
                example: alice
              createdAt:
                type: integer
                format: int64
                description: Account creation timestamp (Unix seconds)
                example: 1704844800
              disabled:
                type: boolean
                description: Whether the user is disabled
                example: false
              admin:
                type: boolean
                description: Whether the user has admin privileges
                example: false
              groups:
                type: array
                items:
                  type: string
                description: Groups the user belongs to
                example: [ developers, editors ]
    Group:
      type: object
      properties:
        name:
          type: string
          description: Group name
          example: developers
        description:
          type: string
          description: Group description
          example: Development team
        members:
          type: array
          items:
            type: string
          description: List of usernames in the group
          example: [ alice, bob ]
        createdAt:
          type: integer
          format: int64
          description: Creation timestamp (Unix seconds)
          example: 1704844800
    CreateGroupRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: Group name
          example: developers
        description:
          type: string
          description: Group description
          example: Development team
        members:
          type: array
          items:
            type: string
          description: Initial list of member usernames
          example: [ alice, bob ]
...
    GroupsListResponse:
      type: object
      properties:
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from POST /v1/auth/login
    APIKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key obtained from POST /v1/auth/api-key